+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 - Show QR scanner video in fullscreen on mobile for onchain transactions
-
+- Android: fix file upload forms in MoonPay
 - Replace the existing BIP69 lexicographical sorting of tx inputs/outputs with a randomized sorting approach
 
 ## 4.41.0